backend/requirements.txt
fastapi
asyncpg
uvicorn
pyjwt
passlib[bcrypt]
openai
youtube-transcript-api
pandas
aiotube
aiohttp
beautifulsoup4
jinja2
requests
email_validator
python-multipart
google-auth
youtube_transcript_api
openpyxl
scrapetube
xlsxwriter
celery
redis
asyncio
pytz
